requirements.txt
# 核心依赖
streamlit>=1.28.0
PyYAML>=6.0

# 数据处理
pandas>=1.5.0